@@ -4,18 +4,18 @@
  * This is a slightly nicer version of './tic-toc.ts' which only works in Node.
  */
 
-export { tic, toc };
+export { tic, toc }
 
-let timingStack: [string | undefined, number][] = [];
+let timingStack: [string | undefined, number][] = []
 
 function tic(label?: string) {
-  if (label) process.stdout.write(`${label}... `);
-  timingStack.push([label, performance.now()]);
+  if (label) process.stdout.write(`${label}... `)
+  timingStack.push([label, performance.now()])
 }
 
 function toc() {
-  let [label, start] = timingStack.pop()!;
-  let time = (performance.now() - start) / 1000;
-  if (label) process.stdout.write(`\r${label}... ${time.toFixed(3)} sec\n`);
-  return time;
+  let [label, start] = timingStack.pop()!
+  let time = (performance.now() - start) / 1000
+  if (label) process.stdout.write(`\r${label}... ${time.toFixed(3)} sec\n`)
+  return time
 }
