@@ -1,9 +1,9 @@
 /** @type {import('postcss-load-config').Config} */
 const config = {
-  plugins: {
-    tailwindcss: {},
-    autoprefixer: {},
-  },
-};
+	plugins: {
+		tailwindcss: {},
+		autoprefixer: {}
+	}
+}
 
-export default config;
+export default config
