@@ -1,55 +1,55 @@
-import type { NextApiRequest, NextApiResponse } from 'next'
-import { PoolMina, PoolMinaDeployProps, MinaTokenHolder, FungibleToken, FungibleTokenAdmin } from "../../../../contracts/build/src/indexmina.js";
-import { AccountUpdate, Mina, PrivateKey, PublicKey, Cache, fetchAccount, UInt64 } from "o1js";
-import fs, { readFileSync } from 'fs'
-import util from 'util';
-import { resolve } from 'path';
-import { readCache } from '../../lib/cache.js';
-import ZkappWorkerClient from './workerClient';
+// import rom 'fs'
+import type { NextApiRequest, NextApiResponse } from "next"
+import {} from "o1js"
+import {} from "../../../../contracts/build/src/indexmina"
+import ZkappWorkerClient from "./workerClient"
 
-export const maxDuration = 300; // This function can run for a maximum of 300 seconds
+export const maxDuration = 300 // This function can run for a maximum of 300 seconds
 
 type ResponseData = {
-    message: string
+	message: string
 }
 
-type SwapData = { user: string, amountIn: number, amountOut: number, balanceOutMin: number, balanceInMax: number };
+type SwapData = {
+	user: string
+	amountIn: number
+	amountOut: number
+	balanceOutMin: number
+	balanceInMax: number
+}
 
-const ZKAPP_ADDRESS = 'B62qrn4bTWsKGddKeLGzriYVQF23fNF4tCnACKawP7ySJfH7zFmd2u6';
-const ZKTOKEN_ADDRESS = 'B62qjDaZ2wDLkFpt7a7eJme6SAJDuc3R3A2j2DRw7VMmJAFahut7e8w';
+const ZKAPP_ADDRESS = "B62qrn4bTWsKGddKeLGzriYVQF23fNF4tCnACKawP7ySJfH7zFmd2u6"
+const ZKTOKEN_ADDRESS = "B62qjDaZ2wDLkFpt7a7eJme6SAJDuc3R3A2j2DRw7VMmJAFahut7e8w"
 
-let inited = false;
-let client;
+let inited = false
+let client
 
 async function init() {
-    if (!inited) {
-        console.log('Loading web worker...');
-        client = new ZkappWorkerClient();
-        await timeout(1);
-        console.log('Done loading web worker');
-        await client.setActiveInstanceToDevnet();
-        await client.loadContract();
-        console.log('Compiling zkApp...');
-        await client.compileContract()
-        console.log('zkApp compiled');
-        inited = true;
-    }
+	if (!inited) {
+		console.log("Loading web worker...")
+		client = new ZkappWorkerClient()
+		await timeout(1)
+		console.log("Done loading web worker")
+		await client.setActiveInstanceToDevnet()
+		await client.loadContract()
+		console.log("Compiling zkApp...")
+		await client.compileContract()
+		console.log("zkApp compiled")
+		inited = true
+	}
 }
 
 async function timeout(seconds: number): Promise<void> {
-    return new Promise<void>((resolve) => {
-        setTimeout(() => {
-            resolve();
-        }, seconds * 1000);
-    });
+	return new Promise<void>((resolve) => {
+		setTimeout(() => {
+			resolve()
+		}, seconds * 1000)
+	})
 }
 
-export default async function handler(
-    req: NextApiRequest,
-    res: NextApiResponse<any>
-) {
-    await init();
+export default async function handler(req: NextApiRequest, res: NextApiResponse<any>) {
+	await init()
 
-    const fib = await client.getReserves("B62qisgt5S7LwrBKEc8wvWNjW7SGTQjMZJTDL2N6FmZSVGrWiNkV21H");
-    return res.status(200).json({ msg: "hello swap", fib: fib });
+	const fib = await client.getReserves("B62qisgt5S7LwrBKEc8wvWNjW7SGTQjMZJTDL2N6FmZSVGrWiNkV21H")
+	return res.status(200).json({ msg: "hello swap", fib: fib })
 }
