@@ -1,23 +1,22 @@
-import { create } from "zustand";
+import { create } from "zustand"
 
 export interface Load {
-  msg: string;
-  state: boolean;
-  process: number;
+	msg: string
+	state: boolean
+	process: number
 }
 
-
 interface LoadState extends Load {
-  update: (value: Partial<Load>) => void;
+	update: (value: Partial<Load>) => void
 }
 
 const useLoad = create<LoadState>((set) => ({
-  msg: "",
-  state: false,
-  process: 0,
-  update: (value: Partial<Load>) => {
-    set((state) => ({ ...state, ...value }));
-  },
-}));
+	msg: "",
+	state: false,
+	process: 0,
+	update: (value: Partial<Load>) => {
+		set((state) => ({ ...state, ...value }))
+	}
+}))
 
-export default useLoad;
+export default useLoad
